Improve error tracking
This would keep track of exactly where the error happened (file and
line), so if someone posts an error, it would be a lot easier to spot
exactly where the error occurred.

Also, it separates errno and VkResult, so I don't have to use awkward
VK_ERROR_* values for errors from pthread or libc.

Signed-off-by: Shahbaz Youssefi <ShabbyX@gmail.com>